+++ lib/Utils/xfuncs.cpp
@@ -371,13 +371,13 @@ bool string_to_bool(const char *s)
 void xsetreuid(uid_t ruid, uid_t euid)
 {
 	if (setreuid(ruid, euid) != 0)
-		perror_msg_and_die("can't set %cid %d", 'u', (int)ruid);
+		perror_msg_and_die("can't set %cid %lu", 'u', (long)ruid);
 }
 
 void xsetregid(gid_t rgid, uid_t egid)
 {
 	if (setregid(rgid, egid) != 0)
-		perror_msg_and_die("can't set %cid %d", 'g', (int)rgid);
+		perror_msg_and_die("can't set %cid %lu", 'g', (long)rgid);
 }
 
 uid_t getuidbyname(const char* login)
